[bug/#35]-refactor-menu-entity
