+++ src/main/java/yuseteam/mealticketsystemwas/domain/admin/entity/Menu.java
@@ -1,4 +1,4 @@
-package yuseteam.mealticketsystemwas.entity;
+package yuseteam.mealticketsystemwas.domain.admin.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import yuseteam.mealticketsystemwas.domain.restaurant.entity.Restaurant;
+import yuseteam.mealticketsystemwas.entity.Ticket;
 
 import java.util.List;
 //하위 - , 가격, 식권 수, 판매 식권 수, 식당, 음식 카테고리, Ticket(Fk)
