+++ src/main/java/yuseteam/mealticketsystemwas/domain/admin/repository/MenuRepository.java
@@ -1,7 +1,7 @@
 package yuseteam.mealticketsystemwas.domain.admin.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import yuseteam.mealticketsystemwas.entity.Menu;
+import yuseteam.mealticketsystemwas.domain.admin.entity.Menu;
 
 public interface MenuRepository extends JpaRepository<Menu,Long> {
 }
