+++ src/main/java/yuseteam/mealticketsystemwas/entity/Ticket.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import yuseteam.mealticketsystemwas.domain.admin.entity.Menu;
 
 import java.time.LocalDateTime;
